pyproject.toml
[project]
name = "dcbot-basic"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "discord-py>=2.7.1",
    "google-api-core>=2.30.0",
    "google-genai>=1.68.0",
    "python-dotenv>=1.2.2",
]
